Added console application selection screen and model/enums folders
