+++ src/com/reservationsystem/room/Main.java
@@ -1,11 +1,12 @@
 package com.reservationsystem.room;
 
+import com.reservationsystem.room.models.Room;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-        ArrayList<Room> rooms = new ArrayList<Room>();
+        ArrayList<Room> rooms = new ArrayList<>();
         rooms.add(new Room(1, 300, 2, 0, "Double",
                 "2 single bed", true, true));
         rooms.add(new Room(2, 200, 2, 1, "Single",
@@ -15,9 +16,23 @@ public class Main {
         rooms.add(new Room(4, 900, 2, 5, "Penthouse",
                 "2x double, 4x single, 2 baby", false, false));
 
-        for (Room room : rooms) {
-            if (room.availability) {
-                System.out.println("Kamer " + room.roomNumber + " is beschikbaar. Deze kamer is een " + room.roomType + ".");
+        while(true) {
+            System.out.println("\nType '1' to list all available rooms.");
+
+            Scanner scanner = new Scanner(System.in);
+            int userInput = Integer.parseInt(scanner.nextLine());
+
+            switch (userInput) {
+                case 1:
+                    for (Room room : rooms) {
+                        if (room.availability) {
+                            System.out.println("Kamer " + room.roomNumber + " is beschikbaar. Deze kamer is een " + room.roomType + ".");
+                        }
+                    }
+                    break;
+                default:
+                    System.out.println("Enter a valid input option!");
+                    break;
             }
         }
     }
